Fix possible crash when no extractable file types are configured in SCFG
If libarchive couldn't extract a QWK or REP packet, we'd fallback to searching
for a match among the configured extractable file types and if no extension/type
match was found, default to the first configured extractable file type
(even if there wasn't one) which would result in a NULL pointer dereference
and most likely a crash. Instead, if no matching configured extractable
file type is found, just log a warning message and don't continue with the
extraction attempt. With SBBS v3.19+, it's totally valid/legit to have no
extractable file types configured in SCFG and things "just work" (using
libarchive).