New config library for dealing with new ctrl/????.ini files
Replaces functionality of cnflib.js for sbbs v3.20+

Only the xtrn.ini read/write support has been added so far
(for use in install-xtrn.js).

I didn't think a lib like this would be needed, being that the
files are now in .ini format, but there was some re-arranging
of property values/section names that still needed to be done by
something (unfortunately), so this does that. But it does make
updating existing scripts that use cnflib.js a breeze.