text_sec updated for v3.20+
There are some changes in here that I forgot the history behind, but the
most important is that it uses file.ini rather than file.cnf, thus
fixing issue #478. I actually had this changed file sitting outside of my
git repo for a while now and didn't realize it.

The other changes appear to be:
- RIP support (clearing the window before/after each file is displayed)
- read the display files a chunk at a time (rather than read the entire file
  into RAM, which could take a really long time on very large files),
- allow files that don't exist on disk to be added to the list