Update U_* variables (pseudo consts) to use v3.20 user.tab field numbers
The first argument to system.matchuserdata() in v3.20 is the user.tab field
number, not the user.dat byte-offset that was used in versions of SBBS prior
to v3.20.

These values are synchronized with userfields.h, but I keept the U_ prefix
here (instead of USER_) to maintain compatibility with existing JS scripts.

This is a fix for issue #501 - thanks for the report Eric!