We need store the file full path in the section name
We need to support the same filename in different directories.

The 'path' key is just an artifact and not normally used when read,
so that was a distraction that apparently confused me previously
and I broke this functionality in commit c44ed79a3b2c5e5.

This should fix the issues reported by xbit (XBITBBS) on DOVE-Net.

Also, don't open the file.ini for write-access (unnecessary).