More details in "Failure to send header/request to FastCGI socket" log msg
Elevate the log level back to ERROR (from WARNING), as this is not an expected
condition and the sysop (me) should be alerted right away. This had previously
be lowered (along with some other log messages) from ERROR to WARNING.

When php-fpm is updated (on Debian, anyway), e.g. from 8.1 to 8.2, a new
etc/php configuration (pool.d) directory was used which set me back to a
default www.conf file that uses Unix domain sockets instead of TCP sockets.