Allow system.matchuserdata() to search deleted user records
Insert an optional boolean argument to enable a search through deleted and
inactive user records (in addition to active user records). This is part of
the solution to the problem described in issue #513 where the UEDIT sysop
command did not go to a specified user (by alias) automatically, presumably
because the user record was marked as deleted.