All UEDIT sysop command to pull up deleted users by alias or number
Requires the latest v3.20a for sytem.matchuserdata() to support deleted and
inactive user record searching.

This is part of the fix for the issue described in issue #513 where the
UEDIT sysop command was falling back (displaying/editing) user #1 because
the sysop-specified user number or alias was not an active user account.

The fallback to display/editing user #1 still happens, but first we try
really hard to find the user that the sysop specified as an argument to
the UEDIT command.