Exports a syncterm.lst file from a FTN nodelist
Paul Hayton (fsxNet admin) had this idea. This seems to do the trick.

It does use the listed IBN (BinkP) address as a fall-back if the INA and ITN
addresses aren't specified, which may not be desireable.
So maybe that should be a command-line option to enable as desired. <shrug>