Add support for an optional description character offset (number) arg
If all descriptions start a fixed offset and the default parsing
logic (regex) isn't working for the sysop, they can specify the exact
error offset to use for the beginning of each file's description. This
offset is only used for the initial line of the description, not the
continuation lines, but perhaps that could be done if needed.

This is an attempt to address issue #530

I first attempted to use "Lookbehind Assertions" in the regex, but
didn't have any success and decided the brute force method that the
old addfiles utility used might be as simpler solution and provide an
effective work-around for more potential issues with auto-detecting
the beginning of the useful file description.

I also added descriptions of the optional arguments to the help output.