Re-write to allow options in any order, more control over all behavior
Previously, in order to just specify the terminal-type string, the sysop
would also have to pass new values for the tg-mode, client-name and
server-name, which was not very friendly. We still support the old syntax
where order of arguments matters, but also a new better syntax for options
(which may now come before or after the required address[:port] argument):

-c <client-name> (default: user alias)
-s <server-name> (default: user real name)
-t <terminal-type> (e.g. "xtrn=doorcode" to auto-exec door on server)
-T <connect-timeout-seconds> (default: 10 seconds)
-m <telnet-gateway-mode> (Number or TG_* vars OR'd together, default: 0)
-p send current user alias and password as server and client-name values
-q don't display banner or pause prompt displayed (quiet)
-P don't pause for user key-press
-C don't clear screen after successful session

For arguments that take a value (e.g. -c, -s, -t, -T, -m), the value may
immediately follow the option letter (e.g. "-cMyName") or be specified
in the following argument (e.g. "-c MyName"). Multiple options cannot be
stuck together in the same option (e.g. use '-C -P' instead of '-CP').

If the RLogin server is a Synchronet BBS, you probably want to specify the
'-p' option which will send the current user's alias and password in the
RLogin connection parameters that Synchronet expects them.

The sysop now has better control over the output (banner, screen-clearing)
and the pause prompt that was previously hard-coded.