Apply inactive_hangup option (for dumb terminals) using max_socket_inactivity
this insures that inactive dumb (bot) connections will be disconnected even
when using a script (e.g. animated pause prompt) that doesn't time-out.

Also, if the connected node is the last node (for this sbbs instance), divide
the socket inactivity timeout value in half.

These changes (along with sbbs v3.20) should help with DOS (denial-of-service,
not MS-DOS) prevention.

Mainly for Krueger.