Support duration notation (e.g. 10M for 10 minutes) in "*Inactivity" .ini keys
The recently updated ctrl/sbbs.ini had MaxLoginInactivity = 10M in
the [BBS] section, which was parsed as 10 seconds rather than the
intended 10 minutes. This fixes that issue and also will now store all
of these inactivity values in duration notation. Existing values stored
in seconds will work fine since that is always the assumed/fallback unit
of time in duration keys.

Thanks for the report!