Add GETDIM @-code to get the current ANSI terminal dimensions (rows and cols)
This code will do nothing if there's no user logged-in or their terminal is
not ANSI or their terminal dimensions (rows and cols) are set to specific
values (not auto-detected). This code does the same thing as calling JS
console.getdimensions(). Since it waits (up to 5 seconds) for a CPR response
from the terminal, this does potentially slow down output, so caveat emptor.

This could be used by a sysop to address issue #529.