Remove the call to sync() from ansi_getlines()
This was needed back in the v2 days to be sure the escape sequence was sent
*after* an output buffer before might've been cleared by a user's abort/Ctrl-C
action. SYNC/ASYNC called riosync() which called rioctl(TXSYNC) and we have no
equivalent in Synchronet for TCP/IP (modern sbbs). A user's Ctrl-C will clear
all pending I/O, but won't prevent subsequent output from being sent (until
the abort condition is cleared) as used to be the case with serial I/O.