Use SMTP-routable Fido/QWK netmail addresses is From header fields
Use the Synchronet-supported FTN/QWKnet routable addressing in the from fields
of messages posted via those network types. Using the mailutil.js
fidoaddr_to_emailaddr() function that was apparently created for listserver.js
but then made unnecessary (and thus unused for several years).

I think this addresses the concerns from issue #487. Reading a message via
NTTP (using the Synchronet NNTP service) in Thunderbird, one can now reply
to sender via email and it "just works" so long as the sysop has configured
the network support for those net-types correctly and Thunderbird is
configured with the Synchronet mail server as its SMTP server to use for that
newsgroup folder. Whether the Synchronet mail server is configured to use
direct delivery or relay is not relevant (though mentioned in the issue
description).