Eliminate node_sem_check and node_stat_check settings
Use the terminal server sem file check interval instead of node_sem_check
(it was only used in the event thread) and node_stat_check was only used in
the control panel, so just support a registry over-ride, but default to 5 sec.

These settings were artifacts from SBBS v2 (WFC mode), and not really relevant
or in the proper configuration place.