By default SBBSecho will now include the 'reason' in bad packet filenames
When bad packets are detected and renamed to ".bad" files, include the reason
in the new filename (e.g. "ff69453a.src-addr.bad" or "ff69453a.pkt-passwd.bad"
instead of just "ff69453a.bad") to make it easier for a sysop to determine
what to do with the bad packets without having to search through log files
to discover the reason *why* SBBSecho considered the packet to be bad.
This behavior can be disabled if desired by setting EchoCfg->Global->Verbose
Bad Packet Names to "No". Note: the reason used is the *last* reason
detected/logged for a packet to be considered bad; it's possible that there
are multiple reasons that a packet may be considered bad. All reasons will be
logged, but only the last reason is used in the new bad packet filename.

Also added the missing "Delete Packets" option (to EchoCfg->Global). Though
this setting was added in sbbsecho v3.0 (and settable via sbbsecho.ini), it
was never exposed via EchoCfg; I don't think that was intentional. Normally,
you'd always want to leave this set to Yes (the default) to delete processed
packets, but when debugging, it may be desirable to leave the packet files
in place. This what the old sbbbsecho v2 '-x' command-line option used to be
used for (disable the deleting of processed packets).

Moved the Global Settings, EchoMail Settings, and File Paths config menus to
the upper left of the screen in EchoCfg. The placement seemed inconsistent
and erratic, so this is a bit of a cosmetic improvement.