Add option to delete bad incoming packets (not rename them)
So a sysop now has 3 choices for SBBSecho's handling of incoming bad packets:
 - delete them
 - rename them *.bad
 - rename them *.reason.bad (the default)

If SBBSecho is configured to rename bad packets, but the rename fails, the
error is logged (as before) but the packet is not deleted (this is a change
from previous behavior).

Always log the *reason* a pkt was determined bad in the "Bad packet detected"
log message.