If ":<port>" suffix is included in an added TCP address, parse/use it
When adding a BBS entry with a TCP connection, if a ":port" suffix is included
(e.g. "vert.synchro.net:23"), parse and use that port number for the new entry
automatically and remove the suffix from the adddress itself.
This is just a time saver when adding new BBS entries.

Care was taken to not treat the following addresses in this manner:
 ":<port>"   - would have resulted in a blank address
 "a::<port>" - possible IPv6 address
 "a:<port>b" - possible IPv6 address or ... ?
 "a:<invalid-port>"