Add TODO comment for cvstat.
When originally created, it was intended to be accessed only by a
single thread, and therefore not need mutex protection, allowing
less code to run with vstatlock held.  Unfortunately, this is not
how it turned out, so cvstat should either be removed, or it should
actually be cleaned up to work as intended.