A basic HTTP-post handler for uploading files to the filebase
Based on qwk.ssjs

I tested this using:
wget --auth-no-challenge --post-file=file.ext --http-user=username --http-password=passwd vert.synchro.net/upload.ssjs?filename=file.ext\?desc=description-text

This currently requires an "uploads" directory to be configured by the sysop.

I'm not sure if there's a better/more-standard way for the posted filename or
content to be included in the request. This is just a sort of proof-of-concept
as a solution for issue #554. Perhaps an ecbwebv4 page displays a prompt to the
user for their filename and it redirects to this page. Not sure how that'll
work.