Push update_pixels down and force a redraw on font change.
update_pixels should only be set when at least one pixel has
actually been changed... setting it when pixels have not been
changed causes unneeded screen updates.

Previously, a font change asserted (incorrectly) that the pixels in
the screen had been modified, so the change would occur at the end
of the next blink cycle when alt fonts are checked.  Now it
requests a full redraw from vmem when a font is changed, which will
trigger a redraw the next time the blinker thread triggers (~5ms).