Fix build on openSUSE Tumbleweed (and building without bitmap)
Some Linux systems ship with a reentrant ncurses library that does
not have an ESCDELAY variable.  For now, simply don't allow control
over ESCDELAY on the systems since we also currently have a global,
and so can't use set_escdelay() as intended.

Also, in ripper, stub out all the stuff that's used when HAS_VSTAT
is not defined.  This makes most of the code not compiled when it's
useless anyway.