Don't link in RGB <-> YUV tables by default.
Instead, have the r2yptr and y2rptr pointers that need to be set
to those tables in order for interpolation to work.  Not setting one
locks ciolib into "LCD" modes (everything uses square pixels), and
XBR scaling is disabled.

The CIOLIB_INTERPOLATE_OBJS make variable contains the path to the
object file that contains those tables, so it's just a matter of
some code and an extra (128MB) object to link against.