Don't treat failure to read node[2+]/node.cnf as a fatal error
There was a period of time when *nix installs of sbbs (after the
migration to Git) didn't include node2/3/4 directories (and thus
the node.cnf files in those directories). In such an install, report
the error converting those node.cnf files to node.ini equivalents,
but don't stop the upgrade to v3.20. The terminal server will fallback
to the firs node config if any other node config (node.ini) file is
missing, so that's still okay (though not ideal).

This fixes issue #560

Thanks for the report Bobrob!