Remove rgbmap.c and rgbmap.h
Instead, generate these as part of the build and use the .incbin
macro on GNU-compatible assemblers to build the .o file quickly.

No good solution for Win32, but genmap could write the C file again
and take forever building it if/when we need to build rgbmap.obj
on Win32.