Perform input translation (e.g. DEL<->BKSPC, PETSCII chars) in input_thread()
My first idea was to have a per-external-program setting to enable input
translation, but Deuce suggested it would be better done in input_thread() and
after some thought, I agree. Translations are not done in data/file transfer
mode, so use the CON_RAW_IN console flag to indicate this condition. So even
JS console.getbyte() will return a translated char unless the console is set
raw input mode (this could be surprising behavior for some!).

I considered saving/restoring the console mode when performing a file transfer
but will leave that to a time when obviously needed.

This fixes issue #497 in automatic way (no option needed).