Prompt the user to chat with The Guru when sysop is not available
This is actually how sbbs v1/2 worked, before this functionality was moved to
chat_sec.src.

If the sysop doesn't want the prompt at all, just set the string
ChatWithGuruInsteadQ to blank in the ctrl/text.dat file (or copy this script
to mods dir and edit to your liking).

This resolves issue #569.