Add options: -utf8, -strip, -user=<name>
The -strip option requires the (new to v3.20) strip_ctrl_a() method

Previously, when using the -all or -lib options, all directories would be
included in the filelist, but with the -user option, a valid user (e.g.
-user=guest, or -user=sysop) can be specfied to limit the scope of the list
to only directories accessible to the specified user (and downloads allowed
to the user as well).

The -utf8 option encodes all output in UTF-8 (instead of CP437) while the
-strip option strips any Ctrl-A (attribute) sequences from the output.