Support narrower (e.g. < 80 column) screen modes better
Maximize screen realestate for options/settings display:
- Don't place a margin around windows when in a narrow screen mode.
- Don't display shadows on windows when in a narrow screen mode.

This commit also fixes a bug in uinput() when left was non-zero and the total
width (left + width) would be too wide for the screen. This bug was apparent in
SCFG->Message Areas in 50 column mode (no Message Group window would be
displayed when a message group was selected).