Add option (for deon) to notify offline users of new received email (SMTP) msgs
The notification of offline users was removed from the mail server 13 years
ago (commit 729eb16b838bbb) without a lot of fanfare or explanation. Add an
option for notification of offline users (in addition to online users) for
those sysops that want that behavior.

Use the userdat is_user_online() function in place of the getnodedat() loop,
which also wasn't considering nodes in "logon" status as having a user online.