Use /bin/pwd instead of pwd to get the PWD.
Some shell pwd implementations default to the "Logical" PWD, which
can contain symlinks.  The POSIX standard for /bin/pwd is to return
the "Physical" PWD with all symlinks resolved.  Some shells don't
support the -P option the the built-in pwd, so we don't want to
reply on that.

For some reason, we're redefining the PWD env variable which should
by the phyical path, but I'm not really interested in tracking down
all the windy history for this.

Basically, this bit me once, and now it's fixed.