Add support for XTerm "Bracketed Paste"
At the same time, add BD, BE, PE, and PS to the terminfo entry.

Note that it seems this is "normally" detected by seeing if $TERM
contains "xterm" which shouldn't work with SyncTERM (which is wildly
incompatible), but the terminfo source file here:
https://invisible-island.net/ncurses/terminfo.ti.html
Gives us hope in the form of this comment:

https://invisible-island.net/xterm/xterm-paste64.html

Bracketed paste was introduced by xterm patch #203 in May 2005, as part of a
larger feature for manipulating the clipboard selection.  Few terminals aside
from xterm fully implement the clipboard feature, but several copy this
detail.  The names for the extended capabilities here were introduced by vim
in January 2017, but used internally.  In 2023, vim patch 9.0.1117 is needed
to work with this change.

That is to say that it likely won't work on anyone's system today
(except maybe Cyan's), but it may magically start working in the
future... assuming tic supports these capnames.  No real clue there
since there's absolutely no termcap support, and I use FreeBSD.