Work around marco or SDL bug in SDL_SetWindowMinimumSize()
When using the marco WM, and resizing using ALT-Right-drag, calls to
SDL_SetWindowMinimumSize() result in the top-left corner of the
window moving up and to the left (appears to be by the border size).

To prevent this from being a maddening issue under marco, ensure
we only call SDL_SetWindowMinimumSize() once when the minimum size
changes on the window.

Fixes SF ticket 115, thanks Ragnarok!

While we're hear, ensure the minimium maximum window size holds
the original sized window... we're not interested in downscaling.