More work on X11 fullscreen mode.
Save the window position and size when we enter fullscreen, and
restore when we leave... this was a bigger box of worms than
expected since we can get ConfigureNotify events that tell is
"Something" changed, but not be able to know the current window
position.

Just remove the x_get_window_info() and anything else that cares
about the window position and derive it from the ALT-Enter keypress
instead.

At the same time, try to remove as much special-casing for
fullscreen as possible.

Tested under marco and xfwm4... will test under Sommelier "soon"