Support inverse mode in atascii
Toggled using backtick.
Current state is shown in the status bar.

Since we previously couldn't actually *show* the status bar in
ATASCII (or C64) mode, enable that as well.  This means supporting
40-column status bars too.

We still default status bar to off for those modes, because that's
likely what people actually want.

Also, it seems ENTER was broken in ATASCII mode, so fix that as well.
Implements SF feature 61