Fix issue with bitmap_drv_init_mode()
Because bitmap_get_scaled_win_size() was being clamped to maxsize,
no matter how much mult was incremented, w and h would never be
larger than maxwidth/maxheight and only under exceptional
circumstances would the be equal, which would result in an infinite
loop.

While we're here, set the integer scaling value so we're not chasing
ulps all over the place with floating point math/scaling when we
don't need to.