Add InactivityAlert, clear out NodeActionMultiChat
The InactivityAlert string is sent as a warning before disconnect when a
visible alert (AreYouThere) shouldn't be sent. Defaults to 3 beeps.

Remove the content of the NodeActionMultiChat (over-rides the hard-coded
default node status) - weird that this was the only node status that reported
the user's security level in the default configuration. This has been this
way since sbbs v2.30, at least. I think it was to be sure that others could
see the age/gender of those they were chatting with. But we have other methods
of reporting that in node status now, so we don't need to use this override
string to accomplish that.