js.exec() expects a full path to the script, normally
There's a weird issue where the scope-walk to find a "js.exec_dir" isn't
working (for Nick Young or Nightfox) after invoking DDMsgReader. This should
fix (or work around) that issue.

And js.exec() doesn't search the mods_dir first when a full script path isn't
passed, so with this change, child scripts that are in the mods_dir should now
work as a sysop would expect.