Support for spying via MQTT between nodes
For read-only spying, the mqtt_spy.js module can be used via jsexec (from an
OS shell command prompt), similar to mosquitto_sub.

For read-write spying, just use the ;SPY sysop command as before and if MQTT
is configured/enabled, it'll use MQTT rather than sockets or shared memory
queues to spy on the target node.

For the first time, you can spy between nodes that are on different servers of
the same BBS using the ;SPY sysop command.

Passes through ctrl characters (except Ctrl-C), unlike the built-in spy
function. This could be revisited later or made optional, but it sems to make
sense to allow Ctrl-Z (e.g. to save a message in fseditor.js) to be passed
through to the target node.